requirements.txt
langchain
langchain-core
langchain_community
python-dotenv

langchain_google_genai

langchain-chroma
langchain_openai

streamlit

